# Use exists subquery to find deleted model instance pks