This is an idea that may improve the performance of `get_deleted` for some of the users in #748.

The changes shift the subquery on the model on `get_deleted` to use the built-in `Exists` (only django versions 1.11 and up are listed as supported as of now, and `Exists` was added in this version). The subquery should work across supported backends, the test suite passes, and I've used locally in MySQL and Postgres successfully.

As far as performance goes, things get a bit tricky, and I've been able to look into query times only in Postgres at the moment. But, in Postgres, I've experienced performance changes similar similar to the following when running `get_deleted` on the entire `Version` queryset:

* For smaller models (in my testing, models with 10,000 or fewer instances), the original query takes 20% of the time of the new query (though on reasonable hardware for models of this size we are talking only a few milliseconds)
* For larger models (in my testing, models with 100,000 or more instances), I've found that the new query takes 20% of the time of the original query (from roughly 10 seconds to 2 seconds in a relatively large model)

But there is another use case in which I've found more significant performance improvements (see jedie/django-reversion-compare#95). In the linked issue, the performance issues are from the following line (https://github.com/jedie/django-reversion-compare/blob/0bfe214e40933e38a5e5a94f3c6d0f56de9051be/reversion_compare/compare.py#L206):

```python
    deleted = list(Version.objects.filter(revision=old_revision).get_deleted(related_model))
```

Here, the original query still must iterate over the pks the related model, but using `Exists` results in a _significant_ performance improvement (in a model with a few million instances and many deletions, the query went from about 50 seconds to 1.5 milliseconds).

That being said, there are a couple of issues:

1. The performance improvement for larger models and filtered querysets may not be worth the cost for other cases
2. Should work with django's natively supported database backends, but performance may vary