@jeremy-engel 2019-05-30
To get some cross-database benchmarks, I used `TestModel` from `test_app`. To populate, the objects were created in a revision, saved in another revision, and then every other object was deleted using the following:

```python
from test_app.models import TestModel
import reversion
from reversion.models import Version


def create_test_objects(num_objects):
    for obj_num in range(num_objects):
        with reversion.create_revision():
            obj = TestModel.objects.create()
        with reversion.create_revision():
            obj.save()
        # Delete half of created objects
        if obj_num % 2 == 0:
            obj.delete() 
```

I ran the original query vs. the version in this pull request for postgres, sqlite, and mysql with 100, 1,000, and 10,000 objects (and 100,000 for postgres) for `Version.objects.get_deleted(TestModel)`:

Postgres 11.3

| Number of objects | Original query | New query |
|------------------:|---------------:|----------:|
| 100               | 1.5 ms         | 1.7 ms    |
| 1000              | 4.8 ms         | 5.8 ms    |
| 10000             | 92.3 ms        | 35.0 ms   |
| 100000            | 933.7 ms       | 257.7 ms  |


MySQL 8.0

| Number of objects | Original query | New query  |
|------------------:|---------------:|-----------:|
| 100               | 2.1 ms         | 7.7 ms     |
| 1000              | 12.8 ms        | 237.7 ms   |
| 10000             | 85.8 ms        | 22313.4 ms |


SQLite 3.24

| Number of objects | Original query | New query |
|------------------:|---------------:|----------:|
| 100               | 167 ms         | 166 ms    |
| 1000              | 1422 ms        | 1398 ms   |
| 10000             | 25430 ms       | 22557 ms  |

So MySQL should get the old fallback subquery again, but the postgres results are improved.