@etianen 2019-06-03
Thanks for a great PR, and a solid set of benchmarks! Merged.